Develop (#17)
* :bug: Remove valid for trainer

* :truck: Reorganize project with local classifier

* :construction: Add train-local

* :bug: Fix train local and calculate score for test

* :construction: Run local model

* :construction: Initial commit to bug fix parse local labels

* :bug: Bug fix local labels

* :construction: load new experiments datasets

* :construction: New local to global labels

* :construction: Clen clode

* Merge branch 'release/0.0.6' of github.com:Sette/hmc_torch into release/0.0.6

* :shirt: Add pre-commit support

* :see_no_evil: Add .pre-commit-config.yaml to .gitignore

* 🙈 Add .python-version to .gitignore

* 🔥 Remove .python-version

* :shirt: Update all procject with black

* :construction: Update train_local and project imports

* :truck: Create notebooks path

* 📝 Update readme

* :truck: Create src path

* :construction: Update func train_local

* :shirt: Make lint in all project

* :rocket: Echo all parameters in train.sh

* :shirt: Using optuna to search parameters

* :truck: Reestructure project to isolate constrained model and baseline models

* :truck: Reestructure local models

* :truck: Reestructure local models

* :truck: Update constrained model class name

* :truck: Update constrained model class name

* 🔧 Update makefile

* 🚨 Lint  in arguments.py with pylama

* :truck: Rename fma dataset notebook name

* :construction: Update train.sh to use hp values

* 🎨 Add new arguments

* ✨ HPO with optuna

* Feature/0.0.6/hp values with args (#4)

* :truck: Reestructure local models

* :truck: Reestructure local models

* :truck: Update constrained model class name

* :truck: Update constrained model class name

* 🔧 Update makefile

* 🚨 Lint  in arguments.py with pylama

* :truck: Rename fma dataset notebook name

* :construction: Update train.sh to use hp values

* 🎨 Add new arguments

* ✨ HPO with optuna

* :bug: Bug fix arguments

* :shirt: Lint some files

* :construction_worker: Update pre commit file

* :construxtion_worker: Update pre commit

* :construction: Add dataset type argument

* :construction: add data to dvc  with gdrive

* :construction: Update to use device for args

* :truck: Reestructure project

* :construction: Add active_levels argument

* :construction: Earling stopping with optuna

* :fire: Remove dvc files

* 📝 Update dataset download docs

* 🙈 Add dataset zip from kaggle to .gitignore

* ♻️ Refactor arguments.py

* ♻️  Refactor train local files

* :shirt: Lint some files

* update train parameters

* :construction: Update to use optuna.TrialPruned

* :zap: New hpo values for level 1

* :truck: Move arguments do root path

* :zap:  Add epochs to eval

* :bug: Fix save test score result to json

* :bug: Verify active_levels args

* :bug: Fix Early-stopping

* :bug: Fix return best_val_loss

* :zap: Reorganize project and add precision_recall_fscore_support for test_step

* ⚡️ Improve performance with active_levels argument

* :construction: Save  precision_recall_fscore_support for test

* :bug: Fix test_step to use dictionaries for local inputs and outputs

* :book: Add docstring to save_dict_to_json for better clarity and usage examples

* :book: Add detailed docstrings for training, validation, and testing steps

* Set threshold = 0.3

* :construction: Update args with hpo values

* remove active_levels in train.sh

* Refactor validation step to check active levels before processing model updates

* Refactor local classifier training and evaluation pipeline

- Changed HMCLocalModel to use nn.ModuleDict instead of nn.ModuleList for levels.
- Created a new main training script for local classifiers, consolidating training, validation, and testing steps.
- Implemented separate validation and testing functions to improve code organization.
- Removed the old train_local.py file and integrated its functionality into the new structure.
- Updated logging to provide more informative outputs during training and evaluation.
- Ensured compatibility with hyperparameter optimization and data preprocessing steps.

* Implement HMCLocalModelHPO class and update training pipeline for hyperparameter optimization

* Refactor hyperparameter optimization to include F1 score and update early stopping patience

* Update hyperparameter values for training configuration

* Enhance hyperparameter optimization and evaluation metrics

- Update early stopping criteria to consider both validation F1 score and loss.
- Refactor validation and testing steps to include precision, recall, and F1 score metrics.
- Improve logging for model evaluation and early stopping conditions.
- Add a new utility function to display metrics for losses and scores.

* Update validation logic to use non-strict inequality for F1 score comparisons

* Refactor validation logic to use strict inequality for F1 score comparisons and update training configuration parameters

* Refactor linter commands in pre-commit configuration and Makefile to use glob patterns for improved file matching; reorganize import statements for consistency across modules.

* Refactor pre-commit configuration to use relative paths for Black and Flake8; update dataset path handling in dataset retrieval functions for consistency.

* Update dataset path handling in get_dataset_paths function for consistency

* Remove unused argument parsing and training scripts to streamline the codebase.

* Refactor optimizer handling in train_step function to simplify code and use a single optimizer instance.

* :construction: Refactor dataset handling and optimizer initialization in training process for improved clarity and consistency

* Add deploy script for remote training and update Python version in pyproject.toml

* Enhance deployment script with git clone and dependency installation options for improved user experience

* Remove default dataset path assignment in train script to allow for dynamic dataset handling

* Comment out local score display in validation step to streamline logging during training

* Refactor dataset path handling in dataset manager and training script for consistency

* Rename HMCLocalModel to HMCLocalConstrainedModel and update references in training script for consistency

* Update logging in dataset manager and adjust target type in validation step for consistency

* Update poetry installation command to exclude root package and adjust target type in validation step for consistency

* Update target type in training step to ensure consistency in loss computation

* Add dataset_path argument to initialize_dataset_experiments in training scripts for consistency

* :construction: Refactor validation and training steps to improve output handling and ensure consistency in loss computation

* :construction: Refactor test_step to simplify output handling by removing constrained output computation

+++ src/hmc/train/local_classifier/baseline/valid.py
@@ -49,8 +49,8 @@ def valid_step(args):
             for index in args.active_levels:
                 if args.level_active[index]:
                     output = outputs[str(index)]
-                    target = targets[index].float()
-                    loss = args.criterions[index](output, target)
+                    target = targets[index]
+                    loss = args.criterions[index](output.double(), target)
                     local_val_losses[index] += loss
 
                     if i == 0:
